refactor : 계산 로직 class 분리, interface 추가
