+++ src/main/java/com/programmers/java/engin/io/Calculation.java
@@ -1,0 +1,5 @@
+package com.programmers.java.engin.io;
+
+public interface Calculation {
+    int getResult(String input);
+}
